# add-fastmcp-to-fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/todos": {
      "get": {
        "summary": "List Todos",
        "description": "List all todos.",
        "operationId": "list_todos",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  },
                  "type": "array",
                  "title": "Response List Todos"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Todo",
        "description": "Create a new todo and return it with its generated id.",
        "operationId": "create_todo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{todo_id}": {
      "put": {
        "summary": "Update Todo",
        "description": "Replace a todo's title and completed state, found by id.",
        "operationId": "update_todo",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Id of the todo to update.",
              "title": "Todo Id"
            },
            "description": "Id of the todo to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Todo",
        "description": "Delete a todo by id.",
        "operationId": "delete_todo",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Id of the todo to delete.",
              "title": "Todo Id"
            },
            "description": "Id of the todo to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Todo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Todo": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short text of the task to do."
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Whether the task is done.",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique id assigned to the todo."
          }
        },
        "type": "object",
        "required": [
          "title",
          "id"
        ],
        "title": "Todo"
      },
      "TodoIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short text of the task to do."
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "Whether the task is done.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}